requirements.txt
selenium==3.141.0
pytest==5.2.1
pytest-html==2.0.0
pytest-rerunfailures==7.0
click==6.7
poium==0.5.1
configparser
Pillow
pytesseract
configparser